+++ include/ArduinoJson/Data/ListConstIterator.hpp
@@ -38,6 +38,14 @@ class ListConstIterator {
     return *this;
   }
 
+  ListConstIterator<T> &operator+=(size_t distance) {
+    while (_node && distance) {
+      _node = _node->next;
+      --distance;
+    }
+    return *this;
+  }
+
  private:
   const ListNode<T> *_node;
 };